2.Airflow_and_S3/requirements.txt
apache-airflow==2.2.5
boto3==1.23.9
pandas==1.4.2
pyarrow==8.0.0
